fix runtime error left shift warning

+++ src/cache.c
@@ -191,7 +191,7 @@ cache_entry_t *cache_add(const char *url)
 
     cache_entry_t *entry = osAllocMem(sizeof(cache_entry_t));
 
-    entry->hash = (sha256_calc[0] << 24) | (sha256_calc[1] << 16) | (sha256_calc[2] << 8) | (sha256_calc[3] << 0);
+    entry->hash = ((uint32_t)sha256_calc[0] << 24) | ((uint32_t)sha256_calc[1] << 16) | ((uint32_t)sha256_calc[2] << 8) | (sha256_calc[3] << 0);
     entry->original_url = strdup(url);
     entry->file_path = custom_asprintf("%s%c%s.%s", cachePath, PATH_SEPARATOR, sha256_calc_str, extension);
     entry->cached_url = custom_asprintf("/cache/%s.%s", sha256_calc_str, extension);
